// src/utils/enums/enums.ts
export enum ErrorCode {
    INTERNAL_ERROR = "INTERNAL_ERROR",
    VALIDATION_FAILED = "VALIDATION_FAILED",
    REQUIRED_TOKEN = "REQUIRED_TOKEN",
    EXPIRED_TOKEN = "EXPIRED_TOKEN",
    INVALID_TOKEN_TYPE = "INVALID_TOKEN_TYPE",
    INVALID_TOKEN_CLAIM = "INVALID_TOKEN_CLAIM",
    INVALID_ROLES_CLAIM = "INVALID_ROLES_CLAIM",
    INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
    INVALID_REQUEST = "INVALID_REQUEST",
    NOT_FOUND_USER = "NOT_FOUND_USER",
    DUPLICATE_USER = "DUPLICATE_USER",
    DELETED_USER = "DELETED_USER",
    MISSING_REQUIRED_DATA = "MISSING_REQUIRED_DATA"
}

export enum SuccessCode {
    OPERATION_SUCCESS = "OPERATION_SUCCESS"
}